requirements.txt
python-ldap
python-dotenv
requests